Move page entry points to page directory
This moves the entry points out of the client root, reducing the
number of folders one has to scan in order to understand what's
happening.
