+++ app/View/Settings.php
@@ -93,7 +93,7 @@ class Settings implements HasActions, HasFilters {
 	 * @since    0.1.0
 	 */
 	public function display_plugin_admin_page() {
-		echo $this->template->render( 'settings/index', $this->get_initial_state() );
+		echo $this->template->render( 'page/settings/index', $this->get_initial_state() );
 	}
 
 	/**
