--- client/editor/instance/delete.hbs
@@ -1,1 +1,0 @@
-{{#> "editor/instance/button"}}<button type="button" data-brk-onclick="onDeleteClick">Delete</button>{{/"editor/instance/button"}}
