--- client/editor/instance/filename.hbs
@@ -1,1 +1,0 @@
-{{#> "editor/instance/button"}}<span contenteditable data-brk-oninput="onFilename">{{instance.filename}}</span>{{/"editor/instance/button"}}
