--- client/editor/instance/toolbar.hbs
@@ -1,1 +1,0 @@
-<div class="toolbar">{{> editor/instance/filename}}{{> editor/instance/language}}{{> editor/instance/delete}}</div>
