+++ client/page/editor/common.scss
