+++ client/page/editor/controls/__tests__/controls.spec.js
