+++ client/page/editor/index.hbs
@@ -1,15 +1,15 @@
 <div data-brk-container="editor" class="wpgp-editor">
     <div class="wpgp-editor-row">
-        {{> editor/description/index}}
+        {{> page/editor/description/index}}
     </div>
 
     <div class="wpgp-editor-row">
-        {{> editor/controls/index}}
+        {{> page/editor/controls/index}}
     </div>
 
     {{#each editor.instances as |instance|}}
         <div class="wpgp-editor-row">
-            {{> editor/instance/index instance=instance}}
+            {{> page/editor/instance/index instance=instance}}
         </div>
     {{/each}}
 </div>
