+++ client/page/editor/index.scss
