+++ client/page/editor/instance/delete.hbs
@@ -1,0 +1,1 @@
+{{#> "page/editor/instance/button"}}<button type="button" data-brk-onclick="onDeleteClick">Delete</button>{{/"page/editor/instance/button"}}
