+++ client/page/editor/instance/filename.hbs
@@ -1,0 +1,1 @@
+{{#> "page/editor/instance/button"}}<span contenteditable data-brk-oninput="onFilename">{{instance.filename}}</span>{{/"page/editor/instance/button"}}
