+++ client/page/editor/instance/index.hbs
@@ -7,5 +7,5 @@
             data-brk-oninput="onInput"
             data-brk-onkeyup="onKeyup"
             data-brk-onkeydown="onKeydown"
-            contenteditable>{{instance.code}}</code>{{> editor/instance/toolbar}}</pre>
+            contenteditable>{{instance.code}}</code>{{> page/editor/instance/toolbar}}</pre>
 </div>
