+++ client/page/editor/instance/index.js
