+++ client/page/editor/instance/language.hbs
@@ -1,5 +1,5 @@
-{{#> "editor/instance/button"}}<select data-brk-onchange="onLanguage">
+{{#> "page/editor/instance/button"}}<select data-brk-onchange="onLanguage">
     {{#each editor.languages}}
         <option value="{{this}}" {{#compare instance.language this}}selected{{/compare}}>{{@key}}</option>
     {{/each}}
-</select>{{/"editor/instance/button"}}
+</select>{{/"page/editor/instance/button"}}
