+++ client/page/editor/instance/toolbar.hbs
@@ -1,0 +1,1 @@
+<div class="toolbar">{{> page/editor/instance/filename}}{{> page/editor/instance/language}}{{> page/editor/instance/delete}}</div>
