+++ client/page/editor/root.js
