+++ client/page/settings/highlighting.hbs
