+++ client/page/settings/index.hbs
@@ -1,15 +1,15 @@
 <div class="wrap" data-brk-container="settings">
-    {{> settings/header}}
+    {{> page/settings/header}}
 
     {{#compare route "bootstrap"}}
         Loading...
     {{/compare}}
 
     {{#compare route "highlighting"}}
-        {{> settings/highlighting}}
+        {{> page/settings/highlighting}}
     {{/compare}}
 
     {{#compare route "accounts"}}
-        {{> settings/accounts}}
+        {{> page/settings/accounts}}
     {{/compare}}
 </div>
