+++ client/page/tinymce/initial.hbs
