+++ client/page/tinymce/root.js
@@ -1,6 +1,6 @@
 // @flow
 import { component, children } from 'brookjs';
-import { SearchComponent } from '../component';
+import { SearchComponent } from '../../component';
 
 export default component({
     children: children({
